refactor(auth): extract JwtConfig + AuthContext into shared epigraph-auth crate
