+++ crates/epigraph-api/src/routes/claims.rs
@@ -1931,7 +1931,8 @@ mod tests {
 #[cfg(all(test, feature = "db"))]
 mod db_tests {
     use super::*;
-    use crate::middleware::bearer::{AuthContext, ClientType};
+    use crate::middleware::bearer::AuthContext;
+    use crate::middleware::ClientType;
     use crate::state::{ApiConfig, AppState};
     use axum::body::Body;
     use axum::http::{Request, StatusCode};
