+++ crates/epigraph-auth/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "epigraph-auth"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+chrono = { workspace = true }
+jsonwebtoken = "9"
+serde = { workspace = true }
+uuid = { workspace = true }
